feat(impact-scores): Add WIP ImactScores screen
This screens comes along with own hook to retreive and parse data, app function to connect to the external API and both ImpactScore screen and Hetamap Chart components.

Heatmap chart is totally reusable (based on some inputs conditions) and it will be more reusable in a future with styling by props, hability to parse more data, etc.